refactor(lock): moved impl to various mods
That way, we improve overall structure and make additions more easy.

Also we added a frame for the `FileLock` type, which is a convenience
wrapper around the existing `Lock` functionality.
